Extend test by `export class` syntax
I stumbled across this issue when working on integration between Next.js and NestJS.
NestJS requires using decorators in every imaginable way and Next.js requires
using Babel instead of TSC, so that's why I this plugin is important for that
usecase.

Apparently, the plugin doesn't play well with the `export class` syntax, which is exposed
by adding the syntax to this test.